hha_prj/requirements.txt
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
docker==5.0.2
docker-compose==1.29.2
PyJWT==2.1.0